set up favorite facility entity, controller, repo, service and add/remove/list endpoints

+++ backend/src/main/java/app/fitnessfinderapp/backend/userAccount/UserAccountRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserAccountRepository extends JpaRepository<UserAccount, Integer> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
   Optional<UserAccount> findByUsername(String username);
 }
